tests/wpt/web-platform-tests/wake-lock/wakelock-supported-by-feature-policy.html

<!DOCTYPE html>
<title>Test that wake-lock is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-feature-policy/#dom-featurepolicy-features">
<link rel="help" href="https://w3c.github.io/wake-lock/#dfn-wake-lock-feature">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('screen-wake-lock', document.featurePolicy.features());
}, 'document.featurePolicy.features should advertise screen-wake-lock.');
</script>
